+++ cli/replicaset/cconfig_test.go
@@ -9,7 +9,6 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/tarantool/tt/cli/replicaset"
-	"github.com/tarantool/tt/cli/running"
 )
 
 var _ replicaset.Discoverer = &replicaset.CConfigInstance{}
@@ -23,13 +22,6 @@ var _ replicaset.Promoter = &replicaset.CConfigApplication{}
 var _ replicaset.Demoter = &replicaset.CConfigApplication{}
 var _ replicaset.Expeller = &replicaset.CConfigApplication{}
 var _ replicaset.VShardBootstrapper = &replicaset.CConfigApplication{}
-
-func TestCConfigApplication_BootstrapVShard(t *testing.T) {
-	app := replicaset.NewCConfigApplication(running.RunningCtx{}, nil, nil)
-	err := app.BootstrapVShard(replicaset.VShardBootstrapCtx{})
-	assert.EqualError(t, err,
-		`bootstrap vshard is not supported for an application by "centralized config" orchestrator`)
-}
 
 func TestCConfigInstance_Discovery(t *testing.T) {
 	cases := []struct {
@@ -456,11 +448,4 @@ func TestCConfigInstance_Expel(t *testing.T) {
 	err := instance.Expel(replicaset.ExpelCtx{})
 	assert.EqualError(t, err,
 		`expel is not supported for a single instance by "centralized config" orchestrator`)
-}
-
-func TestCConfigInstance_BootstrapVShard(t *testing.T) {
-	instance := replicaset.NewCConfigInstance(nil)
-	err := instance.BootstrapVShard(replicaset.VShardBootstrapCtx{})
-	assert.EqualError(t, err, `bootstrap vshard is not supported for a single instance by `+
-		`"centralized config" orchestrator`)
 }